Revise transpose and paragraph commands
Paragraph commands: sa-forward-paragraph, sa-backward-paragraph.  They
check if the last command was sa-transpose-lines, if yes it is
repeated with the same arguments; the appropriate paragraph command is
run otherwise.  The paragraph commands are aware of special Org
paragraph commands (note: they don't take arguments yet).

The transpose lines command stashes the last arguments in a variable
which is retrieved later when transpose commands need to be repeated.